Migrate OSCommand into public API for sst-test (#1284)
* Migrate OSCommand into public API for sst-test

sst_unittest_support.py has an alias from OSCommand to os_command to
make testing sst-elements happy for now.
